Add TestingAttestationBuilder
