+++ eth2/types/src/test_utils/mod.rs
@@ -1,3 +1,6 @@
 mod test_random;
+mod testing_attestation_builder;
 
+pub use rand::{prng::XorShiftRng, SeedableRng};
 pub use test_random::TestRandom;
+pub use testing_attestation_builder::TestingAttestationBuilder;
