+++ eth2/types/src/test_utils/test_random.rs
@@ -1,7 +1,5 @@
 use rand::RngCore;
 
-pub use rand::{prng::XorShiftRng, SeedableRng};
-
 mod address;
 mod aggregate_signature;
 mod bitfield;
